AOSCOS: LOONGARCH: Remove useless caller-saved float registers check
It never successes, because pd_last_callee_saved_fpu_reg is the same as
pd_last_fpu_reg.
LoongArch ABI specified $f24-$f31 as callee-saved registers, while $f31
is the last register. There is no hole between the last callee-saved
float register and the last float register, so the check always
evaluates to false.

Thanks to clang for finding this.

Signed-off-by: Bingwu Zhang <xtex@aosc.io>