AOSCOS: Remove invalid 'd' suffix for double literals
They are causing compilation errors on my Clang setup.
It seems to work with GCC as a GCC extension. No matter how compilers
behave, it is still a GCC extension, not a standard.
According to C++ 14 specification, "2.14.4 Floating literals", 'd' is
not a floating-suffix. Having no suffix
Having no suffix makes a constant of double, which is expected here, so
it can be removed safely.

Signed-off-by: Bingwu Zhang <xtex@aosc.io>